fix postType seed * rename and move the repositories to a repositories folder in every module * organize the services in each module
