+++ src/comments/comments.module.ts
@@ -1,7 +1,7 @@
 import { CommentsController } from "./controllers/comments.controller";
 import { _$ } from "../_domain/injectableTokens";
-import { CommentsService } from "./services/comments.service";
-import { CommentsRepository } from "./services/commentRepository/comments.repository";
+import { CommentsService } from "./services/comments/comments.service";
+import { CommentsRepository } from "./repositories/comment/comments.repository";
 import { DatabaseAccessLayerModule } from "../database-access-layer/database-access-layer.module";
 import { forwardRef, Module } from "@nestjs/common";
 import { HttpModule } from "@nestjs/axios";
