+++ src/neo4j/services/neo4j.service.ts
@@ -39,7 +39,6 @@ export class Neo4jService {
     }
 
     public write(cypher: string, params: Record<string, any>, database?: string): Result {
-        // console.debug(cypher, "cypher", params, "params");
         const session = this.getWriteSession(database);
         return session.run(cypher, params);
     }
