+++ src/posts/models/post.spec.ts
@@ -3,9 +3,9 @@ import { Test, TestingModule } from "@nestjs/testing";
 import { NEO4J_DRIVER, NEO4J_OPTIONS } from "../../neo4j/neo4j.constants";
 import { createDriver } from "../../neo4j/neo4j.utils";
 import { Neo4jService } from "../../neo4j/services/neo4j.service";
-import { PostsRepository } from "../services/postRepository/posts.repository";
+import { PostsRepository } from "../repositories/post/posts.repository";
 import { neo4jCredentials } from "../../_domain/constants";
-import { IPostsRepository } from "../services/postRepository/posts.repository.interface";
+import { IPostsRepository } from "../repositories/post/posts.repository.interface";
 import { Post } from "./post";
 import { _$ } from "../../_domain/injectableTokens";
