+++ src/posts/repositories/post/posts.repository.interface.ts
