+++ src/posts/repositories/postAward/postAward.repository.interface.ts
