+++ src/posts/repositories/postTag/postTags.repository.interface.ts
