+++ src/posts/repositories/postType/postTypes.repository.interface.ts
@@ -3,7 +3,5 @@ import { PostType } from "../../models";
 export interface IPostTypesRepository {
     findAll(): Promise<PostType[]>;
 
-    findPostTypeById(postTypeId: string): Promise<PostType | undefined>;
-
     findPostTypeByName(postTypeName: string): Promise<PostType | undefined>;
 }
