+++ src/posts/services/postReport/postsReport.service.interface.ts
