+++ src/users/models/user.spec.ts
@@ -1,11 +1,11 @@
-import { IUsersRepository } from "../services/usersRepository/users.repository.interface";
+import { IUsersRepository } from "../repositories/users/users.repository.interface";
 import { Test, TestingModule } from "@nestjs/testing";
 import { NEO4J_DRIVER, NEO4J_OPTIONS } from "../../neo4j/neo4j.constants";
 import { Neo4jConfig } from "../../neo4j/neo4jConfig.interface";
 import { neo4jCredentials } from "../../_domain/constants";
 import { createDriver } from "../../neo4j/neo4j.utils";
 import { Neo4jService } from "../../neo4j/services/neo4j.service";
-import { UsersRepository } from "../services/usersRepository/users.repository";
+import { UsersRepository } from "../repositories/users/users.repository";
 import { User } from "./user";
 import { _$ } from "../../_domain/injectableTokens";
